Change to QueryExecutorHolder
